feat: avoiding ifdef, renaming and line ending fix
