+++ etc/config/wazuh-agent.yml
