+++ src/modules/logcollector/src/another_reader_unix.cpp
@@ -1,0 +1,14 @@
+#include <sstream>
+#include <iomanip>
+
+#include <logger.hpp>
+#include <logcollector.hpp>
+
+namespace logcollector {
+
+void AddPlatformSpecificReader([[maybe_unused]] std::shared_ptr<const configuration::ConfigurationParser> configurationParser, [[maybe_unused]] Logcollector &logcollector)
+{
+
+}
+
+}
