+++ src/modules/logcollector/src/event_reader_win.hpp
